Alterar Ocorrencia com Turma

+++ src/SME.SGP.Infra/Dtos/Ocorrencias/Detalhes/OcorrenciaDto.cs
@@ -13,7 +13,7 @@ namespace SME.SGP.Infra
         public bool ConsideraHistorico { get; set; }
         public int AnoLetivo { get; set; }
         public long DreId { get; set; }
-        public int UeId { get; set; }
+        public long UeId { get; set; }
         public int Modalidade { get; set; }
         public int Semestre { get; set; }
         public long? TurmaId { get; set; }
